Improve navigation bar colors
