+++ Apollon/Views/Diagram/DiagramListView.swift
@@ -6,7 +6,6 @@ struct DiagramListView: View {
     @Environment(\.modelContext) private var modelContext
     @Query private var diagrams: [ApollonDiagram]
     @State private var isImporting = false
-    @State private var shouldNavigate = false
     @State private var importErrorMessage: String = ""
 
     var body: some View {
@@ -111,7 +110,6 @@ struct DiagramListView: View {
                 }
             }
             .navigationBarTitleDisplayMode(.inline)
-            .toolbarBackground(ApollonColor.lightGray, for: .navigationBar)
             .toolbarBackground(.visible, for: .navigationBar)
         }
     }
@@ -125,39 +123,5 @@ struct DiagramListView: View {
             }
         }
         return Result { try String(contentsOf: url) }
-    }
-}
-
-struct ImportErrorMessageView: View {
-    @Binding var errorMessage: String
-
-    var body: some View {
-        VStack(alignment: .leading) {
-            HStack {
-                Text("Import failed")
-                    .foregroundColor(.red)
-                    .bold()
-                Spacer()
-                Button {
-                    errorMessage = ""
-                } label: {
-                    Image(systemName: "x.circle")
-                        .foregroundColor(.red)
-                }
-            }
-            Text(errorMessage)
-                .foregroundColor(.red)
-        }
-        .padding(5)
-        .overlay {
-            RoundedRectangle(cornerRadius: 3)
-                .stroke(.red, lineWidth: 1)
-        }
-        .background {
-            RoundedRectangle(cornerRadius: 3)
-                .foregroundColor(Color.red.opacity(0.1))
-        }
-        .padding(.top, 10)
-        .padding(.horizontal, 10)
     }
 }
